Optimize some and Strengthen test
Optimize some and Strengthen test
