+++ 401-react-clickable-grid/401-submit/index.html
@@ -23,6 +23,6 @@
     <!-- develope
     <script src="http://localhost:8080/build/bundle.js" charset="utf-8"></script>
     -->
-    <script src="./build/bundle.js" charset="utf-8"></script>
+    <script src="http://localhost:8080/build/bundle.js" charset="utf-8"></script>
   </body>
 </html>
