+++ 401-react-clickable-grid/401-submit/setup.js
@@ -1,0 +1,7 @@
+import jsdom from 'jsdom';
+
+if (typeof document === 'undefined') {
+  global.document = jsdom.jsdom('<!doctype html><html><body></body></html>');
+  global.window = document.defaultView;
+  global.navigator = global.window.navigator;
+}
