+++ 401-react-clickable-grid/401-submit/src/css/index.scss
@@ -31,44 +31,39 @@ body{
 .main{
   width: 80%;
   margin: auto;
-}
-
-.main ul{
-  padding: 0;
-  display: flex;
-  list-style: none;
-  margin: 10px;
-}
-
-.main ul li:before{
-  content: "";
-  display: inline-block;
-  padding-bottom: 100%;
-  vertical-align: middle;
-}
-
-.main ul li{
-  flex: 1;
-  margin: 0 5px;
-  text-align: center;
-  font-weight: bold;
-  background-color: #f0ad4e;
-  cursor: pointer;
-  opacity: 0.8;
-}
-
-.main ul li:hover{
-  opacity: 1;
-  -webkit-box-shadow:0 0 15px #46b8da;
-  -moz-box-shadow:0 0 15px #46b8da;
-  box-shadow:0 0 10px #46b8da;
+  ul {
+    padding: 0;
+    display: flex;
+    list-style: none;
+    margin: 10px;
+    li:before {
+      content: "";
+      display: inline-block;
+      padding-bottom: 100%;
+      vertical-align: middle;
+    }
+    li {
+      flex: 1;
+      margin: 0 5px;
+      text-align: center;
+      font-weight: bold;
+      background-color: #f0ad4e;
+      cursor: pointer;
+      opacity: 0.8;
+    }
+    li:hover {
+      opacity: 1;
+      -webkit-box-shadow:0 0 15px #46b8da;
+      -moz-box-shadow:0 0 15px #46b8da;
+      box-shadow:0 0 10px #46b8da;
+    }
+  }
 }
 
 .message{
   text-align: center;
   color: red;
-}
-
-.message h3{
-  font-weight: bold;
+  h3{
+    font-weight: bold;
+  }
 }
