+++ 401-react-clickable-grid/401-submit/src/js/grid-view.js
@@ -1,30 +1,5 @@
 import React from 'react';
-
-class GridElement extends React.Component {
-  render() {
-    let rowIndex = this.props.rowIndex;
-    let fontSize = this.props.fontSize;
-    let n = this.props.n;
-    let gridRow = [];
-
-    let style = {
-      fontStyle: {
-        fontSize: fontSize
-      }
-    }
-
-    for(let i = 0; i < n; i++) {
-      let number = rowIndex * n + i + 1;
-      gridRow.push(<li id = {number} key = {'li' + number} style = {style.fontStyle} >{number}</li>);
-    }
-
-    return (
-      <ul key = {'ul' + rowIndex} >
-        {gridRow}
-      </ul>
-    );
-  }
-}
+import GridElement from './gridElement-view';
 
 class GridView extends React.Component {
   render() {
